python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "datasynth-py"
version = "0.1.1"
description = "Python wrapper for DataSynth synthetic data generation"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["synthetic-data", "data-generation", "testing", "machine-learning"]

dependencies = []

[project.optional-dependencies]
cli = ["PyYAML>=6.0"]
memory = ["pandas>=2.0"]
streaming = ["websockets>=12.0"]
all = [
    "PyYAML>=6.0",
    "pandas>=2.0",
    "websockets>=12.0",
]
dev = [
    "PyYAML>=6.0",
    "pandas>=2.0",
    "websockets>=12.0",
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "mypy>=1.0",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://github.com/ey-asu-rnd/RustSyntheticData"
Documentation = "https://github.com/ey-asu-rnd/RustSyntheticData/tree/main/docs"
Repository = "https://github.com/ey-asu-rnd/RustSyntheticData"

[tool.setuptools.packages.find]
where = ["."]
include = ["datasynth_py*"]

[tool.mypy]
python_version = "3.9"
strict = true

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
